ai-server/requirements.txt
fastapi[standard]
uvicorn
anthropic
supabase
python-dotenv
pydantic
matplotlib
plotly
kaleido
Pillow